Add Solution for Task 8.1-k
- add a jsr-303 based validator for priorities
- fix the layout in editPriorities.xhtml

+++ kbe-ue8-1/src/main/java/de/htw_berlin/ai_bachelor/kbe/checklistkbe8/mb/PriorityIntervalMB.java
@@ -28,12 +28,12 @@ public class PriorityIntervalMB {
 		return "editToDoList";
 	}
 
-	public int getMin() {
+	public long getMin() {
 		return  interval.getMin();
 	}
 
-	@Min(value = 1)
-	public int getMax() {
+	@Min(value =  1)
+	public long getMax() {
 		return interval.getMax();
 	}
